requirements.txt
Django==4.2.26
django-auth-ldap==5.2.0
python-ldap==3.4.5
django-bootstrap-v5==1.0.11
reportlab==4.0.7
vobject==0.9.6.1